Factorized the logic drawing of handles.
Nodes, EdgeMovement and EdgeTrigger now have methods to displays theirs handles.
Their editors components OnSceneGUI() have beacome much more manageable.

The next step will be to move the logic to draw the handles, gizmos to a seperate, dedicated file.
This would avoid the game logic to be bloated with display/editor logic.
